CCCT-2534 Delivery Progress Redesign More Tab
[AI] Moved the tab's connectivity listener onto the host, so one watcher both refreshes the tabs and re-runs the sync rather than two overlapping ones.
